requirements.txt
fastapi
uvicorn
newspaper3k
google-cloud-aiplatform
python-dotenv